requirements.txt
anyio==3.6.2
black==22.10.0
certifi==2022.9.24
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
identify==2.5.8
idna==3.4
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.1
platformdirs==2.5.3
pre-commit==2.20.0
pyhumps==3.8.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
virtualenv==20.16.6
